Adapting the service to use ngx toastr
